requirements.txt
Django==4.2
environs==9.5.0
django-ckeditor==6.5.1
pygments